refactor(playground-ui): consolidate tabs components
Remove redundant tab implementations and consolidate to a single
elements/tabs component. Add onClose prop support for closeable tabs.
