+++ packages/playground-ui/src/components/ui/elements/tabs/index.ts
@@ -1,1 +1,4 @@
-export * from './tabs';
+export { Tabs } from './tabs-root';
+export { TabList } from './tabs-list';
+export { Tab } from './tabs-tab';
+export { TabContent } from './tabs-content';
