+++ packages/playground-ui/src/domains/mcps/components/MCPDetail.tsx
@@ -1,6 +1,5 @@
 import { MainContentContent } from '@/components/ui/containers/MainContent';
 import { CopyButton } from '@/components/ui/copy-button';
-import { PlaygroundTabs, Tab, TabContent, TabList } from '@/components/ui/playground-tabs';
 import { Badge } from '@/ds/components/Badge';
 import { Txt } from '@/ds/components/Txt';
 import { FolderIcon, Icon, McpServerIcon } from '@/ds/icons';
