+++ packages/playground-ui/src/index.ts
@@ -11,8 +11,6 @@ export * from './components/ui/containers';
 export * from './components/ui/elements';
 export * from './components/threads';
 export * from './components/ui/entity-header';
-export * from './components/ui/playground-tabs';
-export * from './components/ui/elements';
 export * from './types';
 export * from './ds/components/Badge/index';
 export * from './ds/components/Button/index';
